requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=py37h06a4308_0
blas=1.0=mkl
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py37h06a4308_0
coverage=5.5=py37h27cfd23_2
cudatoolkit=10.2.89=hfd86e86_1
cython=0.29.24=py37h295c915_0
faiss-cpu=1.7.1.post2=pypi_0
freetype=2.10.4=h5ab3b9f_0
grpcio=1.36.1=py37h2157cd5_1
importlib-metadata=3.10.0=py37h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
jpeg=9b=h024ee3a_2
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libarchive-c=3.1=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py37h06a4308_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
olefile=0.46=py37_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
pandas=0.25.1=py37he6710b0_0
pillow=8.3.1=py37h2c7a002_0
pip=21.1.3=py37h06a4308_0
protobuf=3.14.0=py37h2531618_1
python=3.7.10=h12debd9_4
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.5.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
setuptools=52.0.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=1.14.0=py37hf484d3e_0
tk=8.6.10=hbc83047_0
torchvision=0.6.1=py37_cu102
tqdm=4.36.1=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
werkzeug=1.0.1=pyhd3eb1b0_0
wget=3.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
